pyproject.toml
[tool.poetry]
name = "yesand"
version = "0.1.0"
description = ""
authors = ["Will Langdale <willdon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-bootstrap5 = "^24.3"
django-admin-interface = "^0.28.9"
django-bootstrap-icons = "^0.9.0"
django-mathfilters = "^1.0.0"
django-environ = "^0.11.2"
cryptography = "^43.0.3"
django-json-widget = "^2.0.1"
django-ninja = "^1.3.0"
django-treebeard = "^4.7.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.8"
pytest = "^8.3.3"
poethepoet = "^0.29.0"
djlint = "^1.35.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"
exclude = ["yesand/migrations"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "A",
    "DJ",
    "TCH",
    "T20",
    "B",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.django-admin]
settings_module = "config.settings"

[tool.poe.tasks]
_check = "ruff check . --fix"
_format = "ruff format"
_djangoformat = "djlint . --reformat --quiet"
_makemigrations = "poetry run python manage.py makemigrations"
_migrate = "poetry run python manage.py migrate"
_resetmigrations="find yesand/migrations -type f ! -name '__init__.py' -delete"
_resetdb = "rm -rf db.sqlite3"
_bootstrap_admin = "poetry run python manage.py loaddata admin_interface_theme_bootstrap.json"

[tool.poe.tasks.test]
cmd = "pytest -v"
help = "Run tests with pytest"

[tool.poe.tasks.format]
sequence = ["_check", "_format", "_djangoformat"]
help = "Run code formatting checks and fixes with ruff and djLint"

[tool.poe.tasks.migrate]
sequence = ["_makemigrations", "_migrate"]
help = "Run Django database migrations"

[tool.poe.tasks.run]
cmd = "poetry run python manage.py runserver"
help = "Run the Django development server"

[tool.poe.tasks.createsuperuser]
cmd = "poetry run python manage.py createsuperuser --username ${username} --email ${email}"
args = [
  { name = "username", default = "admin" },
  { name = "email", default = "admin@yesand.uk" },
]
help = "Create a Django superuser"

[tool.poe.tasks.collectstatic]
cmd = "poetry run python manage.py collectstatic --noinput"
help = "Collect static files"

[tool.poe.tasks.reset]
sequence = [
  "_resetmigrations", 
  "_resetdb", 
  "_makemigrations",
  "_migrate",
  "_bootstrap_admin", 
  "createsuperuser"
]
help = "Reset the database, migrate, reskin admin, make a new superuser."

[tool.poe.tasks.tree]
cmd = "tree -I '__*'"
help = "Show the project directory structure. Useful for using LLMs to debug."
